add id_wilayah, update fact_populasi to use id_unit_ternak instead id_peternakan

+++ services/db-dwh/ddl/dim_lokasi.sql
@@ -1,6 +1,6 @@
 -- Table
 CREATE TABLE dim_lokasi (
-    id SERIAL,
+    id INT8,
     provinsi VARCHAR(100),
     kabupaten_kota VARCHAR(100),
     kecamatan VARCHAR(100),
